backend/requirements.txt
fastapi
uvicorn 
supabase
tensorflow
numpy
python-dotenv
Pillow